+++ CHANGELOG.md
@@ -15,4 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added django-webtest to the development and test dependencies for test-driven development.
 - Added "coverage" as dependency in dev and test group to generate test coverage reports
 - Added a Django App for static pages, currently serving a dummy index page for the whole project ("Hello World" type).
-- Added a Django App for managing Tournaments. This is the main app for the project.
+- Added a Django App for managing Tournaments. This is the main app for the project
+  - A tournament is defined by a start date and an end date.
+  - Teams are defined by name, club of origin, city of origin and a colour scheme (primary and secondary colour) and have a team leader who is a registered user of the app.
+